+++ scripts/lichessTools-start.js
@@ -1,4 +1,6 @@
-const start = () => window.setTimeout(() => lichessTools.start(window.site),100);
+const start = () => window.setTimeout(() => {
+  lichessTools.start(window.site);
+},100);
 const load = window.site?.load;
 if (load) {
   load.then(start);