+++ tools/additionalGlyphs/tool.js
@@ -22,11 +22,15 @@
       'ro-RO': {
         'options.study': 'Studiu',
         'options.additionalGlyphs': 'Simboluri \u00een plus'
-      }
+      },
+      'zh-TW': {
+        'options.study': '\u7814\u7A76',
+        'options.additionalGlyphs': '\u986F\u793A\u984D\u5916\u5716\u793A',
+      },
     }
 
     isStandardGlyph = (glyph) => {
-      return glyph != '#';
+      return !['#', '\uD83D\uDE10', '\uD83D\uDE22', '\uD83D\uDE41', '\uD83D\uDE10', '\uD83D\uDE42', '\uD83D\uDE01'].includes(glyph);
     }
 
     drawGlyphsDirect = () => {