+++ HISTORY.md
@@ -1,5 +1,14 @@
 History of features added to LiChess Tools in time (reversed order)
 
+17 Sep
+
+- random percent in chapter title
+- eval and clock in exported PGN
+
+16 Sep
+
+- added no spoilers tool
+
 15 Sep
 
 - added support for Firefox